Moves `<GameStartScreen>` to its own folder
