+++ src/web-ui/screens/game-start/rules.css
